Refactored the code structure to match the design proposed to Dr. Bean
