+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module marks the src directory as a Python package.
+"""
