+++ src/metrics/csv_metrics.py
@@ -3,7 +3,6 @@ import logging
 from datetime import datetime
 from pathlib import Path
 from zoneinfo import ZoneInfo
-from quest import step
 import pandas as pd
